+++ src/Sentry/Protocol/Envelopes/Envelope.cs
@@ -27,10 +27,7 @@ public sealed class Envelope : ISerializable, IDisposable
     /// Initializes an instance of <see cref="Envelope"/>.
     /// </summary>
     public Envelope(IReadOnlyDictionary<string, object?> header, IReadOnlyList<EnvelopeItem> items)
-    {
-        Header = header;
-        Items = items;
-    }
+        : this(null, header, items) { }
 
     private Envelope(SentryId? eventId, IReadOnlyDictionary<string, object?> header, IReadOnlyList<EnvelopeItem> items)
     {
@@ -149,8 +146,15 @@ public sealed class Envelope : ISerializable, IDisposable
         // Items
         foreach (var item in Items)
         {
-            await item.SerializeAsync(stream, logger, cancellationToken).ConfigureAwait(false);
-            await stream.WriteNewlineAsync(cancellationToken).ConfigureAwait(false);
+            try
+            {
+                await item.SerializeAsync(stream, logger, cancellationToken).ConfigureAwait(false);
+                await stream.WriteNewlineAsync(cancellationToken).ConfigureAwait(false);
+            }
+            catch (Exception e)
+            {
+                logger?.LogWarning("Failed to serialize envelope item", e);
+            }
         }
     }
 
@@ -167,8 +171,15 @@ public sealed class Envelope : ISerializable, IDisposable
         // Items
         foreach (var item in Items)
         {
-            item.Serialize(stream, logger);
-            stream.WriteNewline();
+            try
+            {
+                item.Serialize(stream, logger);
+                stream.WriteNewline();
+            }
+            catch (Exception e)
+            {
+                logger?.LogWarning("Failed to serialize envelope item", e);
+            }
         }
     }
 
@@ -285,10 +296,16 @@ public sealed class Envelope : ISerializable, IDisposable
             header = CreateHeader(eventId);
         }
 
-        var items = new[]
+        var items = new List<EnvelopeItem>
         {
             EnvelopeItem.FromTransaction(transaction)
         };
+
+        if (transaction.TransactionProfiler is { } profiler)
+        {
+            // Profiler.CollectAsync() may throw in which case the EnvelopeItem won't serialize.
+            items.Add(EnvelopeItem.FromProfileInfo(profiler.CollectAsync(transaction)));
+        }
 
         return new Envelope(eventId, header, items);
     }