+++ src/Sentry/SentrySdk.cs
@@ -71,7 +71,17 @@ public static partial class SentrySdk
     /// <remarks>
     /// If the DSN is not found, the SDK will not change state.
     /// </remarks>
-    /// <returns>An object that should be disposed when the application terminates.</returns>
+    /// <returns>An object that can be disposed to disable the Sentry SDK, if desired.</returns>
+    /// <remarks>
+    /// Disposing the result will flush previously-captured events and disable the SDK.
+    /// In most cases there's no need to dispose the result.  There are only a few exceptions where it makes
+    /// sense to dispose:
+    /// <list type="bullet">
+    /// <item>You have additional work to perform that you don't want Sentry to monitor.</item>
+    /// <item>You have used <see cref="SentryOptionsExtensions.DisableAppDomainProcessExitFlush"/>.</item>
+    /// <item>You are integrating Sentry into an environment that has custom application lifetime events.</item>
+    /// </list>
+    /// </remarks>
     public static IDisposable Init() => Init((string?)null);
 
     /// <summary>
@@ -82,7 +92,17 @@ public static partial class SentrySdk
     /// </remarks>
     /// <seealso href="https://develop.sentry.dev/sdk/overview/#usage-for-end-users"/>
     /// <param name="dsn">The dsn.</param>
-    /// <returns>An object that should be disposed when the application terminates.</returns>
+    /// <returns>An object that can be disposed to disable the Sentry SDK, if desired.</returns>
+    /// <remarks>
+    /// Disposing the result will flush previously-captured events and disable the SDK.
+    /// In most cases there's no need to dispose the result.  There are only a few exceptions where it makes
+    /// sense to dispose:
+    /// <list type="bullet">
+    /// <item>You have additional work to perform that you don't want Sentry to monitor.</item>
+    /// <item>You have used <see cref="SentryOptionsExtensions.DisableAppDomainProcessExitFlush"/>.</item>
+    /// <item>You are integrating Sentry into an environment that has custom application lifetime events.</item>
+    /// </list>
+    /// </remarks>
     public static IDisposable Init(string? dsn) => !Dsn.IsDisabled(dsn)
         ? Init(c => c.Dsn = dsn)
         : DisabledHub.Instance;
@@ -91,7 +111,17 @@ public static partial class SentrySdk
     /// Initializes the SDK with an optional configuration options callback.
     /// </summary>
     /// <param name="configureOptions">The configuration options callback.</param>
-    /// <returns>An object that should be disposed when the application terminates.</returns>
+    /// <returns>An object that can be disposed to disable the Sentry SDK, if desired.</returns>
+    /// <remarks>
+    /// Disposing the result will flush previously-captured events and disable the SDK.
+    /// In most cases there's no need to dispose the result.  There are only a few exceptions where it makes
+    /// sense to dispose:
+    /// <list type="bullet">
+    /// <item>You have additional work to perform that you don't want Sentry to monitor.</item>
+    /// <item>You have used <see cref="SentryOptionsExtensions.DisableAppDomainProcessExitFlush"/>.</item>
+    /// <item>You are integrating Sentry into an environment that has custom application lifetime events.</item>
+    /// </list>
+    /// </remarks>
     public static IDisposable Init(Action<SentryOptions>? configureOptions)
     {
         var options = new SentryOptions();
@@ -107,7 +137,17 @@ public static partial class SentrySdk
     /// <remarks>
     /// Used by integrations which have their own delegates.
     /// </remarks>
-    /// <returns>An object that should be disposed when the application terminates.</returns>
+    /// <returns>An object that can be disposed to disable the Sentry SDK, if desired.</returns>
+    /// <remarks>
+    /// Disposing the result will flush previously-captured events and disable the SDK.
+    /// In most cases there's no need to dispose the result.  There are only a few exceptions where it makes
+    /// sense to dispose:
+    /// <list type="bullet">
+    /// <item>You have additional work to perform that you don't want Sentry to monitor.</item>
+    /// <item>You have used <see cref="SentryOptionsExtensions.DisableAppDomainProcessExitFlush"/>.</item>
+    /// <item>You are integrating Sentry into an environment that has custom application lifetime events.</item>
+    /// </list>
+    /// </remarks>
     [EditorBrowsable(EditorBrowsableState.Never)]
     public static IDisposable Init(SentryOptions options) => UseHub(InitHub(options));
 
@@ -272,18 +312,65 @@ public static partial class SentrySdk
         => CurrentHub.AddBreadcrumb(clock, message, category, type, data, level);
 
     /// <summary>
-    /// Runs the callback with a new scope which gets dropped at the end.
+    /// Runs the callback within a new scope.
     /// </summary>
     /// <remarks>
-    /// Pushes a new scope, runs the callback, pops the scope.
+    /// Pushes a new scope, runs the callback, then pops the scope. Use this when you have significant work to
+    /// perform within an isolated scope.  If you just need to configure scope for a single event, use the overloads
+    /// of CaptureEvent, CaptureMessage and CaptureException that provide a callback to a configurable scope.
     /// </remarks>
     /// <see href="https://docs.sentry.io/platforms/dotnet/enriching-events/scopes/#local-scopes"/>
     /// <param name="scopeCallback">The callback to run with the one time scope.</param>
-    [Obsolete("This method is deprecated in favor of overloads of CaptureEvent, CaptureMessage and CaptureException " +
-              "that provide a callback to a configurable scope.")]
     [DebuggerStepThrough]
     public static void WithScope(Action<Scope> scopeCallback)
         => CurrentHub.WithScope(scopeCallback);
+
+    /// <summary>
+    /// Runs the callback within a new scope.
+    /// </summary>
+    /// <remarks>
+    /// Pushes a new scope, runs the callback, then pops the scope. Use this when you have significant work to
+    /// perform within an isolated scope.  If you just need to configure scope for a single event, use the overloads
+    /// of CaptureEvent, CaptureMessage and CaptureException that provide a callback to a configurable scope.
+    /// </remarks>
+    /// <see href="https://docs.sentry.io/platforms/dotnet/enriching-events/scopes/#local-scopes"/>
+    /// <param name="scopeCallback">The callback to run with the one time scope.</param>
+    /// <returns>The result from the callback.</returns>
+    [DebuggerStepThrough]
+    public static T? WithScope<T>(Func<Scope, T?> scopeCallback)
+        => CurrentHub is IHubEx hub ? hub.WithScope(scopeCallback) : default;
+
+    /// <summary>
+    /// Runs the asynchronous callback within a new scope.
+    /// </summary>
+    /// <remarks>
+    /// Asynchronous version of <see cref="ISentryScopeManager.WithScope"/>.
+    /// Pushes a new scope, runs the callback, then pops the scope. Use this when you have significant work to
+    /// perform within an isolated scope.  If you just need to configure scope for a single event, use the overloads
+    /// of CaptureEvent, CaptureMessage and CaptureException that provide a callback to a configurable scope.
+    /// </remarks>
+    /// <see href="https://docs.sentry.io/platforms/dotnet/enriching-events/scopes/#local-scopes"/>
+    /// <param name="scopeCallback">The callback to run with the one time scope.</param>
+    /// <returns>An async task to await the callback.</returns>
+    [DebuggerStepThrough]
+    public static Task WithScopeAsync(Func<Scope, Task> scopeCallback)
+        => CurrentHub is IHubEx hub ? hub.WithScopeAsync(scopeCallback) : Task.CompletedTask;
+
+    /// <summary>
+    /// Runs the asynchronous callback within a new scope.
+    /// </summary>
+    /// <remarks>
+    /// Asynchronous version of <see cref="ISentryScopeManager.WithScope"/>.
+    /// Pushes a new scope, runs the callback, then pops the scope. Use this when you have significant work to
+    /// perform within an isolated scope.  If you just need to configure scope for a single event, use the overloads
+    /// of CaptureEvent, CaptureMessage and CaptureException that provide a callback to a configurable scope.
+    /// </remarks>
+    /// <see href="https://docs.sentry.io/platforms/dotnet/enriching-events/scopes/#local-scopes"/>
+    /// <param name="scopeCallback">The callback to run with the one time scope.</param>
+    /// <returns>An async task to await the result of the callback.</returns>
+    [DebuggerStepThrough]
+    public static Task<T?> WithScopeAsync<T>(Func<Scope, Task<T?>> scopeCallback)
+        => CurrentHub is IHubEx hub ? hub.WithScopeAsync(scopeCallback) : Task.FromResult(default(T));
 
     /// <summary>
     /// Configures the scope through the callback.
@@ -321,6 +408,11 @@ public static partial class SentrySdk
     [EditorBrowsable(EditorBrowsableState.Never)]
     public static SentryId CaptureEvent(SentryEvent evt, Scope? scope)
         => CurrentHub.CaptureEvent(evt, scope);
+
+    internal static SentryId CaptureEventInternal(SentryEvent evt, Scope? scope)
+        => CurrentHub is IHubEx hub
+            ? hub.CaptureEventInternal(evt, scope)
+            : CurrentHub.CaptureEvent(evt, scope);
 
     /// <summary>
     /// Captures an event with a configurable scope.
@@ -404,7 +496,12 @@ public static partial class SentrySdk
     /// <summary>
     /// Captures a transaction.
     /// </summary>
+    /// <remarks>
+    /// Note: this method is NOT meant to be called from user code!
+    /// Instead, call <see cref="ISpan.Finish()"/> on the transaction.
+    /// </remarks>
     [DebuggerStepThrough]
+    [EditorBrowsable(EditorBrowsableState.Never)]
     public static void CaptureTransaction(Transaction transaction)
         => CurrentHub.CaptureTransaction(transaction);