+++ test/Sentry.Profiling.Tests/Properties/AssemblyInfo.cs
@@ -1,0 +1,1 @@
+[assembly: SuppressMessage("Style", "IDE1006:Naming Styles", Justification = "Test project")]