+++ test/Sentry.Tests/Extensibility/DisabledHubTests.cs
@@ -31,6 +31,15 @@ public class DisabledHubTests
     public void WithScope_NoOp() => DisabledHub.Instance.WithScope(null!);
 
     [Fact]
+    public void WithScopeT_NoOp() => DisabledHub.Instance.WithScope<bool>(null!);
+
+    [Fact]
+    public Task WithScopeAsync_NoOp() => DisabledHub.Instance.WithScopeAsync(null!);
+
+    [Fact]
+    public Task WithScopeAsyncT_NoOp() => DisabledHub.Instance.WithScopeAsync<bool>(null!);
+
+    [Fact]
     public void BindClient_NoOp() => DisabledHub.Instance.BindClient(null!);
 
     [Fact]